requirements.txt
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
coverage==7.2.0
cryptography==39.0.1
dependency-injector==4.41.0
Deprecated==1.2.13
dnspython==2.3.0
Flask==2.2.3
Flask-Limiter==3.2.0
Flask-WTF==1.1.1
hupper==1.11
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.2.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
mdurl==0.1.2
mongoengine==0.26.0
mongomock==4.1.2
ordered-set==4.1.0
packaging==23.0
pycparser==2.21
Pygments==2.14.0
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.28.2
rich==13.3.1
sentinels==1.0.0
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.14.1
WTForms==3.0.1
